script now checks existance
made the script a little better, so now it actually checks to see if the file exists before attempting to copy it

i also added the xorg touchpad config to it

+++ getupdates.sh
@@ -12,13 +12,42 @@ bashp_dest=~/.dotfiles/arch/.bash_profile
 xres_source=~/.Xresources
 xres_dest=~/.dotfiles/arch/.Xresources
 
-echo copying $i3_source to $i3_dest
-cp $i3_source $i3_dest
-echo copying $sway_source to $sway_dest
-cp $sway_source $sway_dest
-echo copying $bashp_source to $bashp_dest
-cp $bashp_source $bashp_dest
-echo copying $xres_source to $xres_dest
-cp $xres_source $xres_dest
+touchpad_source=/etc/X11/xorg.conf.d/70-synaptics.conf
+touchpad_dest=~/.dotfiles/arch/70-synaptics.conf
+
+if [ -e $i3_source ]; then
+	echo Copying $i3_source to $i3_dest
+	cp $i3_source $i3_dest
+else
+	echo !! Error copying $i3_source to $i3_dest !!
+fi
+
+if [ -e $sway_source ]; then
+	echo Copying $sway_source to $sway_dest
+	cp $sway_source $sway_dest
+else
+	echo !! Error copying $sway_source to $sway_dest !!
+fi
+
+if [ -e $bashp_source ]; then
+	echo Copying $bashp_source to $bashp_dest
+	cp $bashp_source $bashp_dest
+else
+	echo !! Error copying $bashp_source to $bashp_dest !!
+fi
+
+if [ -e $xres_source ]; then
+	echo Copying $xres_source to $xres_dest
+	cp $xres_source $xres_dest
+else
+	echo !! Error copying $xres_source to $xres_dest !!
+fi
+
+if [ -e $touchpad_source ]; then
+	echo Copying $touchpad_source to $touchpad_dest
+	cp $touchpad_source $touchpad_dest
+else
+	echo !! Error copying $touchpad_source to $touchpad_dest !!
+fi
 
 exit 0
